Move compilation related functions from System to mix_helpers
