+++ lib/appsignal/system.ex
@@ -1,30 +1,12 @@
 defmodule Appsignal.SystemBehaviour do
-  @callback priv_dir() :: String.t
   @callback hostname_with_domain() :: String.t | nil
   @callback root?() :: boolean()
   @callback heroku?() :: boolean()
   @callback uid() :: integer | nil
-  @callback agent_platform() :: String.t
 end
 
 defmodule Appsignal.System do
   @behaviour Appsignal.SystemBehaviour
-  @os Application.get_env(:appsignal, :os, :os)
-
-  def priv_dir() do
-    case :code.priv_dir(:appsignal) do
-      {:error, :bad_name} ->
-        # This happens on initial compilation
-        Mix.Tasks.Compile.Erlang.manifests
-        |> List.first
-        |> Path.dirname
-        |> String.trim_trailing(".mix")
-        |> Path.join("priv")
-      path ->
-        path
-        |> List.to_string
-    end
-  end
 
   @doc """
   Get the full host name, including the domain.
@@ -65,43 +47,11 @@ defmodule Appsignal.System do
   # platform, rather than the detected platform in .agent_platform during
   # the diagnose run.
   def installed_agent_architecture do
-    case File.read(Path.join([priv_dir(), "appsignal.architecture"])) do
+    case File.read(Path.join([:code.priv_dir(:appsignal), "appsignal.architecture"])) do
       {:ok, arch} -> arch
       {:error, _} -> nil
     end
   end
 
-  def agent_platform do
-    case force_musl_build?() do
-      true -> "linux-musl"
-      false ->
-        case @os.type do
-          {:unix, :linux} ->
-            agent_platform_by_ldd_version()
-          {_, os} ->
-            to_string(os)
-        end
-    end
-  end
 
-  defp agent_platform_by_ldd_version do
-    try do
-      {output, _} = System.cmd("ldd", ["--version"], stderr_to_stdout: true)
-      case String.contains?(output, "musl") do
-        true -> "linux-musl"
-        false ->
-          ldd_version = List.first(Regex.run(~r/\d+\.\d+/, output))
-          case Version.compare("#{ldd_version}.0", "2.15.0") do
-            :lt -> "linux-musl"
-            _ -> "linux"
-          end
-      end
-    rescue
-      _ -> "linux"
-    end
-  end
-
-  defp force_musl_build? do
-    !is_nil(System.get_env("APPSIGNAL_BUILD_FOR_MUSL"))
-  end
 end
